+++ tests/loggers/test_wandb.py
@@ -13,13 +13,11 @@
 # limitations under the License.
 import os
 import pickle
-import types
-from argparse import ArgumentParser
 from unittest import mock
 
 from pytorch_lightning import Trainer
 from pytorch_lightning.loggers import WandbLogger
-from tests.base import BoringModel
+from tests.base import BoringModel, EvalModelTemplate
 
 
 def get_warnings(recwarn):
@@ -106,6 +104,7 @@ def test_wandb_pickle(wandb, tmpdir):
         """ """
         id = 'the_id'
         step = 0
+        dir = 'wandb'
 
         def project_name(self):
             return 'the_project_name'
@@ -121,6 +120,7 @@ def test_wandb_pickle(wandb, tmpdir):
     )
     # Access the experiment to ensure it's created
     assert trainer.logger.experiment, 'missing experiment'
+    assert trainer.log_dir == logger.save_dir
     pkl_bytes = pickle.dumps(trainer)
     trainer2 = pickle.loads(pkl_bytes)
 
@@ -158,19 +158,14 @@ def test_wandb_logger_dirs_creation(wandb, tmpdir):
     assert not os.listdir(tmpdir)
 
     version = logger.version
-    model = BoringModel()
-    limit_batches = 5
-    trainer = Trainer(
-        default_root_dir=tmpdir,
-        logger=logger,
-        max_epochs=1,
-        limit_train_batches=limit_batches,
-        limit_val_batches=limit_batches,
-    )
+    model = EvalModelTemplate()
+    trainer = Trainer(default_root_dir=tmpdir, logger=logger, max_epochs=1, limit_val_batches=3, log_every_n_steps=1)
+    assert trainer.log_dir == logger.save_dir
     trainer.fit(model)
 
     assert trainer.checkpoint_callback.dirpath == str(tmpdir / 'project' / version / 'checkpoints')
-    assert os.listdir(trainer.checkpoint_callback.dirpath) == [f'epoch=0-step={limit_batches - 1}.ckpt']
+    assert set(os.listdir(trainer.checkpoint_callback.dirpath)) == {'epoch=0-step=9.ckpt'}
+    assert trainer.log_dir == logger.save_dir
 
 
 def test_wandb_sanitize_callable_params(tmpdir):
@@ -201,3 +196,10 @@ def test_wandb_sanitize_callable_params(tmpdir):
     assert params["something"] == "something"
     assert params["wrapper_something"] == "wrapper_something"
     assert params["wrapper_something_wo_name"] == "<lambda>"
+
+
+@mock.patch('pytorch_lightning.loggers.wandb.wandb')
+def test_wandb_logger_offline_log_model(wandb, tmpdir):
+    """ Test that log_model=True raises an error in offline mode """
+    with pytest.raises(MisconfigurationException, match='checkpoints cannot be uploaded in offline mode'):
+        logger = WandbLogger(save_dir=str(tmpdir), offline=True, log_model=True)